changed some unroll tests

+++ src/test/java/de/monticore/lang/monticar/emadl/IntegrationTest.java
@@ -137,16 +137,17 @@ public abstract class IntegrationTest extends AbstractSymtabTest {
 
     @Test
     public void testUnroll() {
-        Log.getFindings().clear();
-        String[] args = {"-m", "src/test/resources/models", "-r", "instanceTestUnroll.MainRNN", "-b", this.backend};
-        EMADLGeneratorCli.main(args);
-        final Path instanceTestCifarHasFile
-                = Paths.get("./target/generated-sources-emadl/instanceTestUnroll/RNNencdec.training_hash");
-        try {
-            Files.delete(instanceTestCifarHasFile);
-        }
-        catch(Exception e) {
-            assertFalse("Could not delete hash file", true);
+        if(this.backend.equals("GLUON")) {
+            Log.getFindings().clear();
+            String[] args = {"-m", "src/test/resources/models", "-r", "instanceTestUnroll.MainRNN", "-b", this.backend};
+            EMADLGeneratorCli.main(args);
+            final Path instanceTestCifarHasFile
+                    = Paths.get("./target/generated-sources-emadl/instanceTestUnroll/RNNencdec.training_hash");
+            try {
+                Files.delete(instanceTestCifarHasFile);
+            } catch (Exception e) {
+                assertFalse("Could not delete hash file", true);
+            }
         }
     }
